feat: ignore __typename for maxAliases
The intend here is to prevent bad actors from multiplying the resource use of executing a query.

The current implementation is rather simple. It just counts all aliases. But aliases are sometimes used in ways that have no performance overhead.

A good example is renaming output. See here https://github.com/Escape-Technologies/graphql-armor/issues/166

If an alias is used but the field being aliased is unique for the query it probably shouldn't count as an alias.

Another example of a case that is known to be low impact is aliasing __typename. The server knows the type name and simply has to add it to the output.

This recently hit us at Stellate as we started aliasing our __typename insertions to make it clearer what was added by Stellate.